Attempt to fix a bug in the Standard Errors calculated by the back-end.  Also changes return values when hessians and standard errors are requested but cannot be calculated.
git-svn-id: http://openmx.psyc.virginia.edu/svn/trunk@1136 df83fd30-4cca-4c36-ab9d-e8b5583ccbd1

+++ src/omxObjective.h
@@ -90,7 +90,7 @@ struct omxObjective {					// An objective
 	void omxObjectivePrint(omxObjective *source, char* d);					// Pretty-print a (small) matrix
 	
 	/* Helper functions */
-	void omxCalculateStdErrorFromHessian(int scale, omxObjective *oo);		// Does what it says
+	void omxCalculateStdErrorFromHessian(double scale, omxObjective *oo);	// Does what it says
 	
 	/* Helpers related to objective initialization */
 	omxMatrix* omxNewMatrixFromIndexSlot(SEXP rObj, omxState* state, char* const slotName);	// Gets a matrix from an R SEXP slot
